requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
python-dotenv==1.1.0
Werkzeug==3.1.3
supabase==2.15.3
gunicorn==21.2.0